move regex into utils
